requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.27.0
click>=8.1.0
rich>=13.0.0
python-multipart>=0.0.7
sqlite-utils>=3.30
